Move the getX method to the utility file.
